services/logging/requirements.txt
fastapi
uvicorn
kafka-python
elasticsearch
pydantic